+++ lib/galaxy/model/migrate/versions/0117_add_user_activation.py
@@ -23,10 +23,10 @@ def upgrade(migrate_engine):
     # Add the active and activation_token columns to the user table in one try because the depend on each other.
     try:
         user_table = Table( "galaxy_user", metadata, autoload=True )
-        user_active_column.create( table=user_table, populate_default=True)
         user_activation_token_column.create( table=user_table )
-        assert user_active_column is user_table.c.active
         assert user_activation_token_column is user_table.c.activation_token
+        user_active_column.create( table=user_table, populate_default=True)
+        assert user_active_column is user_table.c.active
     except Exception:
         log.exception("Adding columns 'active' and 'activation_token' to galaxy_user table failed.")
 
@@ -39,9 +39,11 @@ def downgrade(migrate_engine):
     # Drop the user table's active and activation_token columns in one try because the depend on each other.
     try:
         user_table = Table( "galaxy_user", metadata, autoload=True )
-        user_active = user_table.c.active
+        # SQLAlchemy Migrate has a bug when dropping a boolean column in SQLite
+        if migrate_engine.name != 'sqlite':
+            user_active = user_table.c.active
+            user_active.drop()
         user_activation_token = user_table.c.activation_token
-        user_active.drop()
         user_activation_token.drop()
     except Exception:
         log.exception("Dropping 'active' and 'activation_token' columns from galaxy_user table failed.")
